Feat: Search input handling for moving search result page
