+++ src/app/_components/(main)/home/search/[keyword]/clientComponents.tsx
@@ -1,0 +1,2 @@
+import Image from "next/image";
+import superNaturalImage from "@/public/images/supernatural.jpg";
