+++ src/app/_components/(main)/home/search/[keyword]/serverComponents.tsx
@@ -1,0 +1,1 @@
+import "server-only";
